Выполнил модуляцию
