+++ index.html
@@ -229,6 +229,9 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-<script src="js/main.js"></script>
+<script src="js/main.js" type="module"></script>
+<script src="js/util.js" type="module"></script>
+<script src="js/data.js" type="module"></script>
+<script src="js/backend.js" type="module"></script>
 </body>
 </html>
